+++ CHANGES.md
@@ -1,5 +1,7 @@
 # ???
 
-_Released ??? ?, ???_
+_Released ?? ??, ????_
 
+- Entry `uuid`s are now generated using `uuid.uuidv7`
+- Previously marked dead entries can now be optionally included in the whole ring linkrot check
 - ???